pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "checkinbox"
version = "0.0.0"
authors = [
  { name="David Timber", email="dxdt@dev.snart.me" },
]
description = "connect to the MTA and check if the email account exists"
dependencies = ["dnspython"]
readme = "README.md"
requires-python = ">= 3.11"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Communications :: Email"
]

[project.urls]
Homepage = "https://github.com/dxdxdt/checkinbox/"
Issues = "https://github.com/dxdxdt/checkinbox/issues"
